🚚 [projects] Rename module `loadtemplate` to `template`
