+++ src/cutty/projects/template.py
