+++ src/cutty/services/create.py
@@ -4,8 +4,8 @@ from collections.abc import Sequence
 from typing import Optional
 
 from cutty.projects.generate import generate
-from cutty.projects.loadtemplate import loadtemplate
 from cutty.projects.repository import ProjectRepository
+from cutty.projects.template import loadtemplate
 from cutty.templates.domain.bindings import Binding
