+++ tests/unit/projects/test_link.py
@@ -4,10 +4,10 @@ import dataclasses
 import pytest
 
 from cutty.projects.common import GenerateProject
-from cutty.projects.loadtemplate import TemplateMetadata
 from cutty.projects.repository import LATEST_BRANCH
 from cutty.projects.repository import ProjectRepository
 from cutty.projects.repository import UPDATE_BRANCH
+from cutty.projects.template import TemplateMetadata
 from cutty.util.git import Repository
 from tests.util.git import updatefile
